Code Refactoring to manage categories in generic way
