+++ app/lib/features/categories/providers/categories_providers.dart
@@ -1,0 +1,17 @@
+import 'package:acter/common/providers/space_providers.dart';
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+enum CategoriesFor { spaces, chats, pins }
+
+typedef CategoriesInfo = ({String spaceId, CategoriesFor categoriesFor});
+
+final categoriesProvider = FutureProvider.family<Categories, CategoriesInfo>(
+    (ref, categoryInfo) async {
+  final maybeSpace =
+      await ref.watch(maybeSpaceProvider(categoryInfo.spaceId).future);
+  if (maybeSpace != null) {
+    return maybeSpace.categories(categoryInfo.categoriesFor.name);
+  }
+  throw 'Space not found';
+});
