+++ app/lib/features/categories/widgets/category_header_view.dart
