+++ app/lib/features/spaces/pages/sub_spaces.dart
@@ -4,8 +4,9 @@ import 'package:acter/common/providers/space_providers.dart';
 import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/widgets/acter_icon_picker/model/acter_icons.dart';
 import 'package:acter/common/widgets/spaces/space_card.dart';
-import 'package:acter/features/spaces/pages/sub-spaces/category_header_view.dart';
-import 'package:acter/features/spaces/pages/sub-spaces/draggable_space_list.dart';
+import 'package:acter/features/categories/draggable_category_list.dart';
+import 'package:acter/features/categories/providers/categories_providers.dart';
+import 'package:acter/features/categories/widgets/category_header_view.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
 import 'package:atlas_icons/atlas_icons.dart';
 import 'package:flutter/material.dart';
@@ -17,91 +18,43 @@ import 'package:phosphor_flutter/phosphor_flutter.dart';
 
 final _log = Logger('a3::space::sub_spaces');
 
-class SubSpaces extends ConsumerStatefulWidget {
+class SubSpaces extends ConsumerWidget {
   static const moreOptionKey = Key('sub-spaces-more-actions');
   static const createSubspaceKey = Key('sub-spaces-more-create-subspace');
   static const linkSubspaceKey = Key('sub-spaces-more-link-subspace');
-
   final String spaceId;
 
   const SubSpaces({super.key, required this.spaceId});
 
   @override
-  ConsumerState<SubSpaces> createState() => _SubSpacesState();
-}
-
-class _SubSpacesState extends ConsumerState<SubSpaces> {
-  final ValueNotifier<bool> showOrganizedView = ValueNotifier(false);
-
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          _buildAppBarUI(),
-          Expanded(child: _buildSubSpacesUI()),
-        ],
-      ),
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      appBar: _buildAppBarUI(context, ref),
+      body: _buildSubSpacesUI(context, ref),
     );
   }
 
-  Widget _buildAppBarUI() {
-    return ValueListenableBuilder(
-      valueListenable: showOrganizedView,
-      builder: (context, value, child) {
-        return value ? organizedAppBarUI() : normalAppBarUI();
-      },
-    );
-  }
-
-  Widget organizedAppBarUI() {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20),
-      child: Row(
-        children: [
-          Text(
-            L10n.of(context).organized,
-            style: Theme.of(context).textTheme.titleLarge,
-          ),
-          const Spacer(),
-          IconButton(
-            icon: Icon(PhosphorIcons.x()),
-            onPressed: () => showOrganizedView.value = false,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget normalAppBarUI() {
-    final spaceName =
-        ref.watch(roomDisplayNameProvider(widget.spaceId)).valueOrNull;
-    final membership = ref.watch(roomMembershipProvider(widget.spaceId));
+  AppBar _buildAppBarUI(BuildContext context, WidgetRef ref) {
+    final spaceName = ref.watch(roomDisplayNameProvider(spaceId)).valueOrNull;
+    final membership = ref.watch(roomMembershipProvider(spaceId));
     bool canLinkSpace =
         membership.valueOrNull?.canString('CanLinkSpaces') == true;
-    return Row(
-      children: [
-        IconButton(
-          icon: Icon(PhosphorIcons.arrowLeft()),
-          onPressed: () => Navigator.pop(context),
-        ),
-        const SizedBox(width: 4),
-        Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(L10n.of(context).spaces),
-            Text(
-              '($spaceName)',
-              overflow: TextOverflow.ellipsis,
-              style: Theme.of(context).textTheme.labelLarge,
-            ),
-          ],
-        ),
-        const Spacer(),
+    return AppBar(
+      title: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(L10n.of(context).spaces),
+          Text(
+            '($spaceName)',
+            overflow: TextOverflow.ellipsis,
+            style: Theme.of(context).textTheme.labelLarge,
+          ),
+        ],
+      ),
+      actions: [
         IconButton(
           icon: Icon(PhosphorIcons.arrowsClockwise()),
-          onPressed: () => addDummyData(),
+          onPressed: () => addDummyData(ref),
         ),
         if (canLinkSpace) _buildMenuOptions(context),
       ],
@@ -118,7 +71,7 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
           key: SubSpaces.createSubspaceKey,
           onTap: () => context.pushNamed(
             Routes.createSpace.name,
-            queryParameters: {'parentSpaceId': widget.spaceId},
+            queryParameters: {'parentSpaceId': spaceId},
           ),
           child: Row(
             children: <Widget>[
@@ -132,7 +85,7 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
           key: SubSpaces.linkSubspaceKey,
           onTap: () => context.pushNamed(
             Routes.linkSubspace.name,
-            pathParameters: {'spaceId': widget.spaceId},
+            pathParameters: {'spaceId': spaceId},
           ),
           child: Row(
             children: <Widget>[
@@ -145,7 +98,7 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
         PopupMenuItem(
           onTap: () => context.pushNamed(
             Routes.linkRecommended.name,
-            pathParameters: {'spaceId': widget.spaceId},
+            pathParameters: {'spaceId': spaceId},
           ),
           child: Row(
             children: [
@@ -156,7 +109,15 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
           ),
         ),
         PopupMenuItem(
-          onTap: () => showOrganizedView.value = !showOrganizedView.value,
+          onTap: () {
+            showDialog(
+              context: context,
+              builder: (context) => DraggableCategoryList(
+                spaceId: spaceId,
+                categoriesFor: CategoriesFor.spaces,
+              ),
+            );
+          },
           child: Row(
             children: [
               Icon(PhosphorIcons.dotsSixVertical()),
@@ -169,24 +130,21 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
     );
   }
 
-  Widget _buildSubSpacesUI() {
-    final spaceCategories = ref.watch(spaceCategoriesProvider(widget.spaceId));
+  Widget _buildSubSpacesUI(BuildContext context, WidgetRef ref) {
+    final spaceCategories = ref.watch(
+      categoriesProvider(
+        (spaceId: spaceId, categoriesFor: CategoriesFor.spaces),
+      ),
+    );
     return spaceCategories.when(
       data: (categories) {
         final List<Category> categoryList = categories.categories().toList();
-        return ValueListenableBuilder(
-          valueListenable: showOrganizedView,
-          builder: (context, value, child) {
-            return value
-                ? DraggableSpaceList(categoryList: categoryList)
-                : ListView.builder(
-                    scrollDirection: Axis.vertical,
-                    shrinkWrap: true,
-                    itemCount: categoryList.length,
-                    itemBuilder: (BuildContext context, int index) {
-                      return _buildCategoriesList(categoryList[index]);
-                    },
-                  );
+        return ListView.builder(
+          scrollDirection: Axis.vertical,
+          shrinkWrap: true,
+          itemCount: categoryList.length,
+          itemBuilder: (BuildContext context, int index) {
+            return _buildCategoriesList(context, categoryList[index]);
           },
         );
       },
@@ -198,7 +156,7 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
     );
   }
 
-  Widget _buildCategoriesList(Category category) {
+  Widget _buildCategoriesList(BuildContext context, Category category) {
     final entries = category.entries().map((s) => s.toDartString()).toList();
     return Card(
       child: ExpansionTile(
@@ -217,9 +175,8 @@ class _SubSpacesState extends ConsumerState<SubSpaces> {
     );
   }
 
-  Future<void> addDummyData() async {
-    final maybeSpace =
-        await ref.watch(maybeSpaceProvider(widget.spaceId).future);
+  Future<void> addDummyData(WidgetRef ref) async {
+    final maybeSpace = await ref.watch(maybeSpaceProvider(spaceId).future);
     if (maybeSpace != null) {
       final categoriesManager = await maybeSpace.categories('spaces');
