+++ app/lib/router/shell_routers/home_shell_router.dart
@@ -27,7 +27,7 @@ import 'package:acter/features/settings/pages/sessions_page.dart';
 import 'package:acter/features/space/pages/space_details_page.dart';
 import 'package:acter/features/space/settings/pages/visibility_accessibility_page.dart';
 import 'package:acter/features/space/settings/widgets/space_settings_menu.dart';
-import 'package:acter/features/spaces/pages/sub-spaces/sub_spaces.dart';
+import 'package:acter/features/spaces/pages/sub_spaces.dart';
 import 'package:acter/features/super_invites/pages/super_invites.dart';
 import 'package:acter/features/space/pages/chats_page.dart';
 import 'package:acter/features/space/pages/members_page.dart';
